Split go server into multiple files and reformat
